fix(phase7): resolve self-review findings

+++ crates/ttd-protocol-rs/lib.rs
@@ -189,12 +189,12 @@ pub struct ObligationState {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct StepResult {
     pub result: StepResultKind,
-    pub tick: i32,
+    pub tick: u64,
 }
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Snapshot {
     pub worldlineId: String,
-    pub tick: i32,
+    pub tick: u64,
 }
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ComplianceModel {
